added cursor headers

+++ ags/object/ags_cursor.h
@@ -45,7 +45,7 @@ struct _AgsCursorInterface
   guint (*cursor_get_rate)(AgsCursor *cursor);
 
   gdouble (*cursor_get_delay)(AgsCursor *cursor);
-  guint (*cursor_get_duration)(AgsCursor *cursor);
+  guint64 (*cursor_get_duration)(AgsCursor *cursor);
 
   gdouble (*get_delay_counter)(AgsCursor *cursor);
   guint64 (*get_offset)(AgsCursor *cursor);
@@ -67,7 +67,7 @@ gdouble ags_cursor_cursor_get_bpm(AgsCursor *cursor);
 guint ags_cursor_cursor_get_rate(AgsCursor *cursor);
 
 gdouble ags_cursor_cursor_get_delay(AgsCursor *cursor);
-guint ags_cursor_cursor_get_duration(AgsCursor *cursor);
+guint64 ags_cursor_cursor_get_duration(AgsCursor *cursor);
 
 gdouble ags_cursor_get_delay_counter(AgsCursor *cursor);
 guint64 ags_cursor_get_offset(AgsCursor *cursor);
